requirements.txt
geopandas~=1.0.1
numpy==1.26.4
xarray==0.20.1
Django~=5.0.7
geojson~=3.1.0
rasterio~=1.3.10
netCDF4~=1.7.1.post1
pyproj~=3.6.1
pandas~=2.0.3
shapely~=2.0.4
affine~=2.4.0
scipy~=1.14.0
matplotlib~=3.9.1
branca~=0.7.2
folium~=0.17.0
APScheduler~=3.10.4
pytesseract~=0.3.10
pillow~=10.4.0